backend/requirements.txt
# Requirements pour l'API et le traitement principal
flask==2.0.3
flask-cors==3.0.10
mysql-connector-python==8.0.28
scikit-learn==1.0.2
joblib==1.1.0

# Deep Learning (DQN)
tensorflow==2.8.0
numpy==1.22.4

# Utilitaires supplémentaires
requests==2.27.1
python-dotenv==0.19.2